requirements.txt
fastapi
uvicorn
scikit-learn
sentence-transformers
faiss-cpu
numpy
pandas
tqdm
joblib
umap-learn
matplotlib
seaborn